Correction to solve problems with loops and % string operator
Due to the changes in visit_BinOp, sometime naturals where derived for
the numbers. To avoid this problem, a check has been added for
integer_types and a long(-1) is returned to ensure the use of integers.
On the other side, a check on the mod operator has been added to see if
the left-side operand is a string.