Bugs aj1s and issue_13.
Issue_13 can be solved just adding the _nrbits in the xor bits.
The bug aj1s can be solved just doing the oposite.
The source of the problem is the necessity to fix the size of the intbv
in case of bit operators, while it is necessary free it when
calculating arithmetic operations between integer numbers and intbv.
A compromise must be found and included in intbv, like truncating the
upper bits of the integer number, in order to solve both bugs.