Compatibilize visit_BinOp and visitUnary with bitarray
To have equivalent behaviors for both, intbv based objects and bitarray
ones, the visit_BinOp has be modified.
Also a check has been included in case of integer values operation a
check to ensure that the result is always a long.
Finally, as the TrueDiv is supported in VHDL, the test has been
eliminated from NotSupported.